+++ Source/IGListKit/Internal/IGListAdapterProxy.m
@@ -29,6 +29,7 @@ static BOOL isInterceptedSelector(SEL sel) {
             sel == @selector(collectionView:didDeselectItemAtIndexPath:) ||
             sel == @selector(collectionView:didHighlightItemAtIndexPath:) ||
             sel == @selector(collectionView:didUnhighlightItemAtIndexPath:) ||
+            sel == @selector(collectionView:contextMenuConfigurationForItemAtIndexPath:point:) ||
             // UICollectionViewDelegateFlowLayout
             sel == @selector(collectionView:layout:sizeForItemAtIndexPath:) ||
             sel == @selector(collectionView:layout:insetForSectionAtIndex:) ||